+++ CHANGELOG.md
@@ -1,26 +1,30 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
 ### Breaking
+
 - `import.file` is removed. Use `import.path` instead for files and directories.
 - `keycloak.migrationKey` is removed. Use `import.cache-key` instead.
 - `keycloak.realm` is removed. Use `import.login-realm` to define the realm to login.
 - If you have defined requiredActions in your realm configure, make sure you have defined all in your json files. All not defined
   actions will removed now by keycloak-config-cli.
 
-
 ### Added
+
 - PMD for static source code analysis
-- *Experimental* GraalVM support. Run keycloak-config-cli without Java!
+- _Experimental_ GraalVM support. Run keycloak-config-cli without Java!
 - Throw errors on unknown properties in config files
 - Add, update and remove clientScopes
 - Removed required actions if they not defined in import json.
 
 ### Changed
+
 - Reduce docker image size
 - Bump SpringBoot from 2.2.7 to 2.3.0
 - Bump keycloak from 10.0.0 to 10.0.2
@@ -34,67 +38,88 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 ### Removed
+
 - `import.file` parameter
 
 ## [1.4.0] - 2020-04-30
+
 ### Added
+
 - AuthenticatorConfig support (thanks JanisPlots)
 - Keycloak 10 support
 
 ### Changed
+
 - Bump keycloak 9.0.3
 
 ### Fixed
+
 - Fix spotbugs and sonar findings
 
 ## [1.3.1] - 2020-04-02
 
 ### Changed
+
 - Bump Spring Boot version to 2.2.5
 - Bump maven-javadoc-plugin from 3.1.1 to 3.2.0
 
 ### Fixed
+
 - Use username filter for updating users, too.
 
 ## [1.3.0] - 2020-03-27
+
 ### Added
+
 - Add and update groups
 - Update composites in roles
 
 ### Changed
+
 - Add copyright header to all java classes
 - Bump Keycloak to 9.0.2
 
 ## [1.2.0] - 2020-03-15
+
 ### Added
+
 - Implement migrationKey property for different config files per realm
 - Implement identity providers
 
 ### Changed
+
 - Add @SuppressWarnings("unchecked")
 - Migrate to maven single module
 - Use TestContainers
 
 ### Fixed
+
 - Correct username on import
 
 ## [1.1.2] - 2020-02-25
+
 ### Changed
+
 - Use Java 8 inside container again
 
 ## [1.1.1] - 2020-02-25
+
 ### Fixed
+
 - Re-add Keycloak 8
 
 ## [1.1.0] - 2020-02-25
+
 ### Added
+
 - Keycloak 9 support
 
 ### Changed
+
 - Use Java 11 inside container
 - Bump hibernate-validator from 6.0.13.Final to 6.1.0.Final
 
-[Unreleased]: https://github.com/adorsys/keycloak-config-cli/compare/v1.4.1...HEAD
+[unreleased]: https://github.com/adorsys/keycloak-config-cli/compare/v1.4.1...HEAD
 [1.4.0]: https://github.com/adorsys/keycloak-config-cli/compare/v1.3.1...v1.4.0
 [1.3.1]: https://github.com/adorsys/keycloak-config-cli/compare/v1.3.0...v1.3.1
 [1.3.0]: https://github.com/adorsys/keycloak-config-cli/compare/v1.2.0...v1.3.0